Add `ICommandHandler::ConnectorUpdateShort()` as a way to distinguish using a short ID string vs. the full connector ID. Adds associated distinct class and a common base class for both ConnectorUpdate* derivatives.
